resolve lint error
